app/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.9
aiofiles==23.2.1
databricks-sdk>=0.20.0
scikit-learn==1.3.2
numpy>=1.24.0
pandas>=2.0.0
joblib>=1.3.0
mlflow>=2.10.0
